Separated Month and Data structs into different files
